back/requirements.txt
flask
flask-cors
pymongo
PyMuPDF
spacy
scikit-learn
Werkzeug
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
